Add train.py, tuning.py, eval.py
